Update CSS Variables and Remove Unused Components
This PR includes the following changes:

CSS Updates: Modified the CSS variables in application.tailwind.css to update color values and add new variables for card and popover.
Component Cleanup: Removed unused components installation_instructions_coming_soon.rb and premium_badge.rb.
New Components: Added new components tailwind_config.rb and tailwind_css.rb.
Documentation Updates: Updated rails_bundler_view.rb and rails_importmaps_view.rb in the installation documentation.
Configuration Update: Modified tailwind.config.js to reflect the new changes.
