+++ app/views/docs/installation/rails_importmaps_view.rb
@@ -73,7 +73,11 @@ class Docs::Installation::RailsImportmapsView < ApplicationView
         steps.add_step do
           step_container do
             TypographyLarge { "Install PhlexUI gem" }
-            phlex_ui_installation("free")
+            TypographyP { "Run the following in the terminal to install the PhlexUI Component Library" }
+            code = <<~CODE
+              bundle add phlex_ui
+            CODE
+            Codeblock(code, syntax: :javascript)
           end
         end
         # STEP 3
@@ -116,28 +120,14 @@ class Docs::Installation::RailsImportmapsView < ApplicationView
         # STEP 2
         steps.add_step do
           step_container do
-            TypographyLarge { "Update Tailwind Configuration" }
-            TypographyP do
-              plain "Add the following to your "
-              TypographyInlineCode(class: "whitespace-nowrap") { "tailwind.config.js" }
-              plain " file"
-            end
-
-            Codeblock(tailwind_config, syntax: :javascript)
+            render Docs::TailwindConfig.new
           end
         end
 
         # STEP 3
         steps.add_step do
           step_container do
-            TypographyLarge { "Add CSS variables" }
-            TypographyP do
-              plain "Add the following to your "
-              TypographyInlineCode { "app/assets/stylesheets/application.tailwind.css" }
-              plain " file"
-            end
-            code = css_variables
-            Codeblock(code, syntax: :css)
+            render Docs::TailwindCss.new
           end
         end
       end
@@ -150,39 +140,6 @@ class Docs::Installation::RailsImportmapsView < ApplicationView
     div(class: "space-y-4", &)
   end
 
-  def phlex_ui_installation(plan)
-    case plan
-    when "free"
-      TypographyP { "Run the following in the terminal to install the PhlexUI Component Library" }
-      code = <<~CODE
-        bundle add phlex_ui
-      CODE
-      Codeblock(code, syntax: :javascript)
-    when "pro"
-      TypographyP do
-        plain "Your license key is "
-        TypographyInlineCode(class: "whitespace-nowrap") { @phlex_ui_pro_private_key }
-      end
-      TypographyP { "Use it to set your license key in development" }
-      code = <<~CODE
-        export BUNDLE_PHLEXUI__FURY__SITE=#{@phlex_ui_pro_private_key}
-      CODE
-      Codeblock(code, syntax: :javascript)
-      TypographyP { "Check that the key is set" }
-      code = <<~CODE
-        echo $BUNDLE_PHLEXUI__FURY__SITE
-      CODE
-      Codeblock(code, syntax: :javascript)
-      TypographyP { "If you see your key printed in the terminal, you're good to continue" }
-
-      TypographyP { "Now you can install the Pro version of PhlexUI" }
-      code = <<~CODE
-        bundle add phlex_ui
-      CODE
-      Codeblock(code, syntax: :javascript)
-    end
-  end
-
   def js_installation
     render Steps::Builder.new do |steps|
       # STEP 1
@@ -191,7 +148,7 @@ class Docs::Installation::RailsImportmapsView < ApplicationView
           TypographyLarge { "Install package" }
           TypographyP { "Run the following in the terminal to install PhlexUI JS package" }
           code = <<~CODE
-            bin/importmap pin phlex_ui
+            bin/importmap pin rbui-js
           CODE
           Codeblock(code, syntax: :javascript)
         end
@@ -206,268 +163,11 @@ class Docs::Installation::RailsImportmapsView < ApplicationView
           plain " file"
         end
         code = <<~CODE
-          import 'phlex_ui';
+          import 'rbui-js';
         CODE
         Codeblock(code, syntax: :javascript)
       end
     end
-  end
-
-  def upgrade_to_pro
-    div(class: "overflow-hidden rounded-lg") do
-      AspectRatio(aspect_ratio: "3/2") do
-        img(
-          alt: "Placeholder",
-          loading: "lazy",
-          src:
-            helpers.image_path("installation_blur.jpg"),
-          class: "opacity-20 blur-xl"
-        )
-        div(class: "absolute inset-0 p-6 flex flex-col items-center justify-center") do
-          Card(class: "max-w-md mx-auto") do
-            CardHeader do
-              CardTitle { "Upgrade to PhlexUI Pro" }
-              CardDescription do
-                "Get access to all components, boost your productivity and take your projects to the next level."
-              end
-            end
-            CardFooter(class: "flex justify-end gap-x-2") do
-              Link(href: helpers.root_path(anchor: :pricing), variant: :primary) do
-                plain "Get all access"
-                arrow_icon
-              end
-            end
-          end
-        end
-      end
-    end
-  end
-
-  def tailwind_config
-    <<~CODE
-      // For importing tailwind styles from phlex_ui/phlex_ui_pro gem
-      const execSync = require('child_process').execSync;
-
-      // Import phlex_ui gem path (To make sure Tailwind loads classes used by phlex_ui gem)
-      const outputPhlexUI = execSync('bundle show phlex_ui', { encoding: 'utf-8' });
-      const phlex_ui_path = outputPhlexUI.trim() + '/**/*.rb';
-
-      const defaultTheme = require('tailwindcss/defaultTheme')
-
-      module.exports = {
-        darkMode: ["class"],
-        content: [
-          './app/views/**/*.{erb,haml,html,slim,rb}',
-          './app/helpers/**/*.rb',
-          './app/assets/stylesheets/**/*.css',
-          './app/javascript/**/*.js',
-          phlex_ui_path
-        ],
-        theme: {
-          container: {
-            center: true,
-            padding: "2rem",
-            screens: {
-              "2xl": "1400px",
-            },
-          },
-          extend: {
-            colors: {
-              border: "hsl(var(--border))",
-              input: "hsl(var(--input))",
-              ring: "hsl(var(--ring))",
-              background: "hsl(var(--background))",
-              foreground: "hsl(var(--foreground))",
-              primary: {
-                DEFAULT: "hsl(var(--primary))",
-                foreground: "hsl(var(--primary-foreground))",
-              },
-              secondary: {
-                DEFAULT: "hsl(var(--secondary))",
-                foreground: "hsl(var(--secondary-foreground))",
-              },
-              destructive: {
-                DEFAULT: "hsl(var(--destructive))",
-                foreground: "hsl(var(--destructive-foreground))",
-              },
-              warning: {
-                DEFAULT: "hsl(var(--warning))",
-                foreground: "hsl(var(--warning-foreground))",
-              },
-              success: {
-                DEFAULT: "hsl(var(--success))",
-                foreground: "hsl(var(--success-foreground))",
-              },
-              muted: {
-                DEFAULT: "hsl(var(--muted))",
-                foreground: "hsl(var(--muted-foreground))",
-              },
-              accent: {
-                DEFAULT: "hsl(var(--accent))",
-                foreground: "hsl(var(--accent-foreground))",
-              },
-            },
-            borderRadius: {
-              lg: `var(--radius)`,
-              md: `calc(var(--radius) - 2px)`,
-              sm: "calc(var(--radius) - 4px)",
-            },
-            fontFamily: {
-              sans: defaultTheme.fontFamily.sans,
-            },
-          },
-        },
-      }
-    CODE
-  end
-
-  def tailwind_config_pro
-    <<~CODE
-      // For importing tailwind styles from phlex_ui/phlex_ui_pro gem
-      const execSync = require('child_process').execSync;
-
-      // Import phlex_ui gem path (To make sure Tailwind loads classes used by phlex_ui gem)
-      const outputPhlexUI = execSync('bundle show phlex_ui', { encoding: 'utf-8' });
-      const phlex_ui_path = outputPhlexUI.trim() + '/**/*.rb';
-
-      // Import phlex_ui_pro gem path (To make sure Tailwind loads classes used by phlex_ui_pro gem)
-      const outputPhlexUIPro = execSync('bundle show phlex_ui_pro', { encoding: 'utf-8' });
-      const phlex_ui_pro_path = outputPhlexUIPro.trim() + '/**/*.rb';
-
-      const defaultTheme = require('tailwindcss/defaultTheme')
-
-      module.exports = {
-        darkMode: ["class"],
-        content: [
-          './app/views/**/*.{erb,haml,html,slim,rb}',
-          './app/helpers/**/*.rb',
-          './app/assets/stylesheets/**/*.css',
-          './app/javascript/**/*.js',
-          phlex_ui_path,
-          phlex_ui_pro_path
-        ],
-        theme: {
-          container: {
-            center: true,
-            padding: "2rem",
-            screens: {
-              "2xl": "1400px",
-            },
-          },
-          extend: {
-            colors: {
-              border: "hsl(var(--border))",
-              input: "hsl(var(--input))",
-              ring: "hsl(var(--ring))",
-              background: "hsl(var(--background))",
-              foreground: "hsl(var(--foreground))",
-              primary: {
-                DEFAULT: "hsl(var(--primary))",
-                foreground: "hsl(var(--primary-foreground))",
-              },
-              secondary: {
-                DEFAULT: "hsl(var(--secondary))",
-                foreground: "hsl(var(--secondary-foreground))",
-              },
-              destructive: {
-                DEFAULT: "hsl(var(--destructive))",
-                foreground: "hsl(var(--destructive-foreground))",
-              },
-              warning: {
-                DEFAULT: "hsl(var(--warning))",
-                foreground: "hsl(var(--warning-foreground))",
-              },
-              success: {
-                DEFAULT: "hsl(var(--success))",
-                foreground: "hsl(var(--success-foreground))",
-              },
-              muted: {
-                DEFAULT: "hsl(var(--muted))",
-                foreground: "hsl(var(--muted-foreground))",
-              },
-              accent: {
-                DEFAULT: "hsl(var(--accent))",
-                foreground: "hsl(var(--accent-foreground))",
-              },
-            },
-            borderRadius: {
-              lg: `var(--radius)`,
-              md: `calc(var(--radius) - 2px)`,
-              sm: "calc(var(--radius) - 4px)",
-            },
-            fontFamily: {
-              sans: defaultTheme.fontFamily.sans,
-            },
-          },
-        },
-      }
-    CODE
-  end
-
-  def css_variables
-    <<~CODE
-      @tailwind base;
-      @tailwind components;
-      @tailwind utilities;
-
-
-      @layer base {
-        :root {
-          --background: 0 0% 100%;
-          --foreground: 0 0% 3.9%;
-          --primary: 0 0% 9%;
-          --primary-foreground: 0 0% 98%;
-          --secondary: 0 0% 96.1%;
-          --secondary-foreground: 0 0% 9%;
-          --muted: 0 0% 96.1%;
-          --muted-foreground: 0 0% 45.1%;
-          --accent: 0 0% 96.1%;
-          --accent-foreground: 0 0% 9%;
-          --destructive: 350 89% 60%;
-          --destructive-foreground: 0 0% 100%;
-          --warning: 38 92% 50%;
-          --warning-foreground: 0 0% 100%;
-          --success: 87 100% 37%;
-          --success-foreground: 0 0% 100%;
-          --border: 0 0% 89.8%;
-          --input: 0 0% 89.8%;
-          --ring: 0 0% 3.9%;
-          --radius: 0.5rem;
-        }
-      
-        .dark {
-          --background: 0 0% 3.9%;
-          --foreground: 0 0% 98%;
-          --primary: 0 0% 98%;
-          --primary-foreground: 0 0% 9%;
-          --secondary: 0 0% 14.9%;
-          --secondary-foreground: 0 0% 98%;
-          --muted: 0 0% 14.9%;
-          --muted-foreground: 0 0% 63.9%;
-          --accent: 0 0% 14.9%;
-          --accent-foreground: 0 0% 98%;
-          --destructive: 350 89% 60%;
-          --destructive-foreground: 0 0% 100%;
-          --warning: 38 92% 50%;
-          --warning-foreground: 0 0% 100%;
-          --success: 84 81% 44%;
-          --success-foreground: 0 0% 100%;
-          --border: 0 0% 14.9%;
-          --input: 0 0% 14.9%;
-          --ring: 0 0% 83.1%;
-        }
-      }
-
-      @layer base {
-        * {
-          @apply border-border;
-        }
-        body {
-          @apply bg-background text-foreground;
-          font-feature-settings: "rlig" 1, "calt" 1;
-        }
-      }
-    CODE
   end
 
   def info_icon
